refactor(viewer): split runtime mapping gates
